--- infra/docs.sh
@@ -1,8 +1,0 @@
-#!/bin/bash
-
-set -e 
-set -x
-
-npm install --global yarn
-yarn install --frozen-lockfile
-yarn docs
