--- infra/test-all.js
@@ -1,23 +1,0 @@
-const { spawn } = require('child_process');
-const { solanaTestValidator, solanaLogs } = require('./helpers');
-
-process.env.NX_TASKS_RUNNER_DYNAMIC_OUTPUT = false;
-
-async function main() {
-  const [solanaValidatorChild, solanaLogsChild] = await Promise.all([
-    solanaTestValidator,
-    solanaLogs,
-  ]);
-
-  const child = spawn('yarn', ['test'], {
-    stdio: [process.stdin, process.stdout, process.stderr],
-  });
-
-  child.on('close', () => {
-    solanaValidatorChild.kill('SIGTERM');
-    solanaLogsChild.kill('SIGTERM');
-    process.exit(0);
-  });
-}
-
-main().then().catch(console.error);
