+++ packages/cli/tests/unit/collateral.spec.ts
@@ -1,7 +1,7 @@
 import { expect } from 'expect';
 import sinon, { SinonStub } from 'sinon';
 
-import { ChildProccess, spawnValidator } from '../validator';
+import { ChildProccess, spawnValidator } from '../../../validator';
 import { Ctx, runCli, ADDRESS, TX, readCtx } from '../helpers';
 
 describe('collateral', () => {
