+++ packages/validator/accounts/base_registered_mint.json
