+++ packages/validator/accounts/mint_authority.json
