+++ packages/validator/accounts/quote_registered_mint.json
