+++ packages/validator/accounts/taker_base_wallet.json
