+++ packages/validator/accounts/taker_collateral.json
