+++ packages/validator/accounts/taker_collateral_info.json
