+++ packages/validator/accounts/taker_collateral_token.json
