+++ packages/validator/accounts/taker_quote_wallet.json
