+++ packages/validator/programs/psyoptions_european_instrument.so
