+++ packages/validator/programs/spot_instrument.so
